+++ src/main/java/am/nuaca/bim/repository/CompilationRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 /**
  * @author Tigran Sargsyan on 15-Apr-20.
  */
-public interface CompilationRepository extends CrudRepository<Compilation, Long> {
+public interface CompilationRepository extends CrudRepository<Compilation, Integer> {
 
 	@Query(value = "SELECT * FROM compilations WHERE MATCH(title) AGAINST(:text IN BOOLEAN MODE)", nativeQuery = true)
 	List<Compilation> searchByTitlePattern(@Param("text") String text);
